+++ examples/basic-usage.js
@@ -3,19 +3,38 @@ const ArgsAndFlags = require('../index')
 const options = {
   args: [
     {
+      name: 'required-arg',
+      type: 'string',
+      help: 'an argument for saying hello',
+      required: true
+    },
+    {
       name: 'hello',
       type: 'string',
       help: 'an argument for saying hello',
       default: 'hey'
+    },
+    {
+      name: 'integer',
+      type: 'integer',
+      help: 'an integer argument',
+      default: 5
     }
   ],
   flags: [
     {
-      name: 'message',
-      alias: 'm',
+      name: 'toggle',
+      alias: 't',
       type: 'boolean',
       help: 'a boolean argument',
       default: true
+    },
+    {
+      name: 'message',
+      alias: 'm',
+      type: 'string',
+      help: 'a string argument',
+      required: true
     }
   ]
 }
@@ -25,3 +44,4 @@ const { args, flags } = parser.parse(process.argv.slice(2))
 
 console.log('args', args)
 console.log('flags', flags)
+console.log(parser.help())